+++ CHANGELOG.md
@@ -14,6 +14,8 @@ FlatLaf Change Log
 
 - Extras: Added missing export of package
   `com.formdev.flatlaf.extras.components` to Java 9 module descriptor.
+- JIDE Common Layer: Invoke `LookAndFeelFactory.installJideExtension()` when
+  using FlatLaf UI delegates. (issue #230)
 
 
 ## 0.46